Rename: ImageInput, TodoButton 내의 style 경로 수정
